refactor: removed mongol_ prefix in mongoldb model
