+++ src/db/mongoldb/model.rs
@@ -1,15 +1,15 @@
-mod mongol_user;
-mod mongol_chat_info; 
-mod mongol_message; 
-mod mongol_bucket; 
-mod mongol_refresh_token;
-mod mongol_relation;
-mod mongol_chat;
+mod user;
+mod chat_info; 
+mod message; 
+mod bucket; 
+mod refresh_token;
+mod relation;
+mod chat;
 
-pub use mongol_user::*;
-pub use mongol_chat_info::*;
-pub use mongol_message::*;
-pub use mongol_bucket::*; 
-pub use mongol_refresh_token::*;
-pub use mongol_relation::*;
-pub use mongol_chat::*;
+pub use user::*;
+pub use chat_info::*;
+pub use message::*;
+pub use bucket::*; 
+pub use refresh_token::*;
+pub use relation::*;
+pub use chat::*;
