+++ src/db/mongoldb/model/chat_info.rs
